+++ tomcat/src/test/java/org/apache/coyote/http11/ResourceFinderTest.java
@@ -1,5 +1,7 @@
 package org.apache.coyote.http11;
 
+import org.apache.coyote.support.ResourceFinder;
+import org.apache.coyote.support.ResourceNotFoundException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
